requirements.txt
boto3
fastapi
python-dotenv
uvicorn
python-multipart